pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dq-platform"
version = "0.1.0"
description = "API-first data quality monitoring platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Job queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # Database connectors
    "pymysql>=1.1.0",
    "pyodbc>=5.0.0",
    "google-cloud-bigquery>=3.14.0",
    "snowflake-connector-python>=3.6.0",
    "redshift-connector>=2.1.0",
    "duckdb>=1.0.0",
    "oracledb>=2.0.0",
    "databricks-sql-connector>=3.0.0",

    # HTTP client (webhook notifications)
    "httpx>=0.25.0",

    # Great Expectations (check engine)
    "great-expectations>=1.0.0",

    # Security
    "cryptography>=41.0.0",

    # Scheduling
    "croniter>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "testcontainers>=3.7.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dq_platform"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
